fixup??? Ensure index matches head before invoking merge machinery, round N
This is not really meant to be squashed in with that commit. It is
merely a workaround: t3030.35 improperly relies on `qsort()` to be
stable (which MSVC's isn't, and it results in a different code path to
be taken in the rename detection of an ambiguous rename between either
`e` to `a` or `a~Temporary merge branch 2_0` to `a`).

Hopefully this makes the Azure Pipeline of `shears/pu` go green again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>